refactor(rust-api): extract error + serde-default helpers out of lib.rs (sc-8890)
F-088: lib.rs was a 2,530-line grab-bag. Extract the shared `ApiError` type
(status + detail, From<store errors>, IntoResponse) into `error.rs`, and the
~40 `#[serde(default = "...")]` value providers plus `bool_is_false` into
`defaults.rs`. Both are re-exported via `pub(crate) use` so every handler
module's `use super::*` and every `#[serde(default = "default_x")]` string
path keeps resolving unchanged — behavior-preserving.

Fields/methods on `ApiError` are now `pub(crate)` (siblings that constructed
it as a struct literal or called its ctors relied on child-of-crate-root
visibility that no longer applies once it lives in a sibling module).

The `STALE_LORA_UPLOAD_SECONDS` -> `STALE_UPLOAD_SECONDS` rename called out by
F-088 was already landed by sc-8885 (F-083); no further change needed there.

Deferred (higher-risk, not attempted to avoid a big-bang split): full
`server.rs` extraction of Settings/AppState/run and the wholesale removal of
`use module::*` glob imports.
